Track Update Checklist (#32)
* Track Update Checklist

* Link to sections in issue templates, remove old changelog requirements

Two changes:

1. Linking to the actual directories and sections of interest in each issue template. For example:
```
[docs](../blob/main/docs)
[quick start](../#quick-start)
-- alternatively
[quick start](../blob/main/README.md/#quick-start)
```

This assumes all repos are using the `main` (versus `master`) naming convention. Should we make this another repo guideline?

If the directory or section does not exist, this will produce a 404 error, but I assume the updater tool will create a placeholder for all of these.

2. Removed changelog requirements from the `docs` issue template.
Should "missing changelog" be a separate issue template, or is it safe to assume that the updater tool will always create a changelog?